remove a copy of due.py within tests, copy and create script within a fixture
