--- duecredit/tests/envs/stubbed/README.txt
@@ -1,2 +1,0 @@
-A sample script with our stub to verify that it runs without complete
-failure even if there are internal duecredit problems
